+++ internal/cartridge/mbc3.go
@@ -8,8 +8,8 @@ type Mbc3Cartridge struct {
 	BaseCartridge
 	Ram        []byte
 	RamEnabled bool
-	RomBank    uint16
-	RamBank    uint16
+	RomBank    int
+	RamBank    int
 }
 
 func NewMbc3Cartridge(header *CartridgeHeader, data []byte) *Mbc3Cartridge {
@@ -28,10 +28,10 @@ func (c *Mbc3Cartridge) ReadByte(address uint16) uint8 {
 	case 0x0000, 0x1000, 0x2000, 0x3000:
 		return c.BinaryData[address]
 	case 0x4000, 0x5000, 0x6000, 0x7000:
-		return c.BinaryData[c.RomBank*0x4000+(address&0x3fff)]
+		return c.BinaryData[c.RomBank*0x4000+int(address&0x3fff)]
 	case 0xA000, 0xB000:
 		if c.RamEnabled {
-			return c.Ram[c.RamBank*0x2000+(address&0x1fff)]
+			return c.Ram[c.RamBank*0x2000+int(address&0x1fff)]
 		} else {
 			return 0xFF
 		}
@@ -50,25 +50,25 @@ func (c *Mbc3Cartridge) WriteByte(address uint16, value uint8) {
 			return
 		}
 	case 0x2000, 0x3000:
-		c.RomBank = uint16(value)
+		c.RomBank = int(value)
 		if int(c.RomBank)*0x4000 >= len(c.BinaryData) {
-			c.RomBank = uint16(int(c.RomBank) % (len(c.BinaryData) / 0x4000))
+			c.RomBank = (int(c.RomBank) % (len(c.BinaryData) / 0x4000))
 		}
 		if c.RomBank == 0 {
 			c.RomBank = 1
 		}
 	case 0x4000, 0x5000:
-		c.RamBank = uint16(value) & 0x3
+		c.RamBank = int(value) & 0x3
 		if len(c.Ram) <= 0 {
 			c.RamBank = 0
 		} else if int(c.RamBank)*0x2000 >= len(c.Ram) {
-			c.RamBank = uint16(int(c.RamBank) % (len(c.Ram) / 0x2000))
+			c.RamBank = (int(c.RamBank) % (len(c.Ram) / 0x2000))
 		}
 
 	case 0x6000, 0x7000:
 	case 0xA000, 0xB000:
 		if c.RamEnabled {
-			c.Ram[c.RamBank*0x2000+(address&0x1fff)] = value
+			c.Ram[c.RamBank*0x2000+int(address&0x1fff)] = value
 		}
 	default:
 		return
